Alter names of packages so that package names are now lowercased
