--- src/main/java/DAO/ISpotifyPlaylistDAO.java
@@ -1,10 +1,0 @@
-package DAO;
-
-import Model.SpotifyPlaylist;
-
-import java.io.IOException;
-
-public interface ISpotifyPlaylistDAO {
-
-	SpotifyPlaylist getPlaylist(String accessToken, String playlistId) throws IOException;
-}
