--- src/main/java/DAO/TokenDAO.java
@@ -1,12 +1,0 @@
-package DAO;
-
-import Model.Token;
-import Exception.*;
-
-import java.io.IOException;
-
-public interface TokenDAO {
-
-	Token getToken() throws IOException, BadRequestException;
-
-}
