--- src/main/java/Exception/BadRequestException.java
@@ -1,8 +1,0 @@
-package Exception;
-
-public class BadRequestException extends RuntimeException {
-
-	public BadRequestException(String errorMessage) {
-		super(errorMessage);
-	}
-}
