--- src/main/java/Exception/InvalidAccessTokenException.java
@@ -1,8 +1,0 @@
-package Exception;
-
-public class InvalidAccessTokenException extends BadRequestException {
-
-	public InvalidAccessTokenException(String errorMessage) {
-		super(errorMessage);
-	}
-}
