--- src/main/java/Exception/InvalidCredentialsForTokenException.java
@@ -1,8 +1,0 @@
-package Exception;
-
-public class InvalidCredentialsForTokenException extends BadRequestException {
-
-	public InvalidCredentialsForTokenException(String errorMessage) {
-		super(errorMessage);
-	}
-}
