--- src/main/java/Exception/InvalidPlaylistIdException.java
@@ -1,8 +1,0 @@
-package Exception;
-
-public class InvalidPlaylistIdException extends BadRequestException {
-
-	public InvalidPlaylistIdException(String errorMessage) {
-		super(errorMessage);
-	}
-}
