--- src/main/java/Exception/InvalidRequestTokenPathException.java
@@ -1,8 +1,0 @@
-package Exception;
-
-public class InvalidRequestTokenPathException extends BadRequestException {
-
-	public InvalidRequestTokenPathException(String errorMessage) {
-		super(errorMessage);
-	}
-}
