--- src/main/java/Model/Artist.java
@@ -1,16 +1,0 @@
-package Model;
-
-public class Artist {
-
-	private String name;
-
-	public Artist() {}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-}
