--- src/main/java/Model/Item.java
@@ -1,16 +1,0 @@
-package Model;
-
-public class Item {
-
-	private Track track;
-
-	public Item() {}
-
-	public Track getTrack() {
-		return track;
-	}
-
-	public void setTrack(Track track) {
-		this.track = track;
-	}
-}
