--- src/main/java/Model/Token.java
@@ -1,7 +1,0 @@
-package Model;
-
-public interface Token {
-
-	String getAccess_token();
-	void setAccess_token(String access_token);
-}
