--- src/main/java/Model/YoutubeAuthorizationOptions.java
@@ -1,7 +1,0 @@
-package Model;
-
-public class YoutubeAuthorizationOptions {
-
-	public static String API_KEY = "AIzaSyDj56wQymMPvZ8gzKb-6WCfW1oVR6ghXGo";
-
-}
