+++ src/main/java/dao/ISpotifyPlaylistDAO.java
@@ -1,0 +1,10 @@
+package dao;
+
+import model.SpotifyPlaylist;
+
+import java.io.IOException;
+
+public interface ISpotifyPlaylistDAO {
+
+	SpotifyPlaylist getPlaylist(String accessToken, String playlistId) throws IOException;
+}
