+++ src/main/java/dao/TokenDAO.java
@@ -1,0 +1,12 @@
+package dao;
+
+import model.Token;
+import exception.*;
+
+import java.io.IOException;
+
+public interface TokenDAO {
+
+	Token getToken() throws IOException, BadRequestException;
+
+}
