+++ src/main/java/exception/BadRequestException.java
@@ -1,0 +1,8 @@
+package exception;
+
+public class BadRequestException extends RuntimeException {
+
+	public BadRequestException(String errorMessage) {
+		super(errorMessage);
+	}
+}
