+++ src/main/java/exception/InvalidAccessTokenException.java
@@ -1,0 +1,8 @@
+package exception;
+
+public class InvalidAccessTokenException extends BadRequestException {
+
+	public InvalidAccessTokenException(String errorMessage) {
+		super(errorMessage);
+	}
+}
