+++ src/main/java/exception/InvalidCredentialsForTokenException.java
@@ -1,0 +1,8 @@
+package exception;
+
+public class InvalidCredentialsForTokenException extends BadRequestException {
+
+	public InvalidCredentialsForTokenException(String errorMessage) {
+		super(errorMessage);
+	}
+}
