+++ src/main/java/exception/InvalidPlaylistIdException.java
@@ -1,0 +1,8 @@
+package exception;
+
+public class InvalidPlaylistIdException extends BadRequestException {
+
+	public InvalidPlaylistIdException(String errorMessage) {
+		super(errorMessage);
+	}
+}
