+++ src/main/java/exception/InvalidRequestTokenPathException.java
@@ -1,0 +1,8 @@
+package exception;
+
+public class InvalidRequestTokenPathException extends BadRequestException {
+
+	public InvalidRequestTokenPathException(String errorMessage) {
+		super(errorMessage);
+	}
+}
