+++ src/main/java/model/Artist.java
@@ -1,0 +1,16 @@
+package model;
+
+public class Artist {
+
+	private String name;
+
+	public Artist() {}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+}
