+++ src/main/java/model/Item.java
@@ -1,0 +1,16 @@
+package model;
+
+public class Item {
+
+	private Track track;
+
+	public Item() {}
+
+	public Track getTrack() {
+		return track;
+	}
+
+	public void setTrack(Track track) {
+		this.track = track;
+	}
+}
