+++ src/main/java/model/SpotifyAuthorizationOptions.java
@@ -1,0 +1,9 @@
+package model;
+
+public class SpotifyAuthorizationOptions {
+
+	public static String CLIENT_ID = "4299b9c3763b4311b4cffa528525e61c";
+	public static String CLIENT_SECRET = "63842318a1944c2eb815a38a6e978730";
+	public static String TOKEN_URL = "https://accounts.spotify.com/api/token";
+
+}
