+++ src/main/java/model/Token.java
@@ -1,0 +1,7 @@
+package model;
+
+public interface Token {
+
+	String getAccess_token();
+	void setAccess_token(String access_token);
+}
