+++ src/main/java/model/YoutubeAuthorizationOptions.java
@@ -1,0 +1,7 @@
+package model;
+
+public class YoutubeAuthorizationOptions {
+
+	public static String API_KEY = "AIzaSyDj56wQymMPvZ8gzKb-6WCfW1oVR6ghXGo";
+
+}
